Organize the subcommands and their methods
I did this to get a better understanding of what was implemented at the
basic functional level and what was left. I also thought that the
organization made more logical sense. I made the methods public as this
should be able to also be used as a library as well.