Add method to addIdTo a patch
I did this so that I could use it eventually in the rr command and
potentially other places to be able to uniquely identify a patch. This
will facilitate hiding branches behind the command line interface
because I will be able to map branch name to patch uuid and store that
state for later reference.