Add ability to get patch stack id of a commit
I did this so that I could use it to check in request review if the
patch already has an idea, because if it does we don't want to change
that id. If it doesn't then we want to give it an id.

This will also be useful when improving the list functionality to
include information about the patches and if they have already had a
review requested, or if they have changed since they last had review
requested.