Fix rr of top most patch bug
There was a bug where when you would request review for the top most
patch on the stack it would fail because during the process of adding
the patch stack identifier to the patch it tries to cherry pick an empty
set which results an error stopping the process.

So, I simply check before cherry picking the commits that there is a set
of commits to actually be cherry picked. I do this by grabbing the
rev-list of the range to be cherry picked. If it is empty I know there
is no set. If it is NOT empty then I know there is something to be
cherry picked.

ps-id: FCC20BFE-68CB-43C8-91F5-DEF4E424D0B4