Fix prob w/ pub/rr after pubing but not pulling
I did this because in the scenario where the user has published a patch
and has yet to pull we were unable to perform any commands that required
addIdTo() to be called because the references it was using are incorrect
when it this particular scenario.

I resolved this issue by adding a method to get the common ancestor
between two refs and then using that common ancestor as the base for the
addIdTo() method.

ps-id: 96A136A1-3B41-4A51-8F1F-FA474C6EFF5E