Add request review repository cleanse to pull
I did this so that when the user performs a pull command and updates
their local git state we also review the currently stored state for
patch stack and clean it up, removing any request review records that
are no longer needed.

This will help keep all the other operations be more performant because
rather than the state just continuously growing over time as it has in
the past. Now every time the user runs the pull command it goes through
and cleans up the git patch stack state store.

ps-id: D2B7684B-E082-450B-AA1B-FEB7ECB97EEB