Fix bug where shell config output broke git-ps
We did this so that users who had echo statements in their shell
configurations (e.g. .bash_profile, etc.) could successfully use git-ps.
Previously it would fail to run any command because it was unable to
successfully find the git command.

The reason it was failing was because the which() implementation was
reading all of standard output in as the path rather than just grabbing
the last line of standard output as the path. Hence, we modified it to
now grab the last line as the path.

This resolves issue #7.

ps-id: 960C0299-DC86-45EA-B490-BBB1BA85BA20