Refactor `run` to use output & error pipe streaming
I did this because we found that when dealing with commands that have
lots of output it would actually cause it to hang because the pipe
buffer limit would be hit and therefore it wouldn't exit the
waitUntilExit call.

This changes the implementation so that it sets up streaming handlers
early on before the waitUntilExit so that while the process is happening
the streaming handlers are receiving the data and collecting it so that
when the application does exit we have it all without impeding commands
that output a lot.

This also has the benefit of improving the performance of the command
execution.

[changelog]
fixed: issue where commands with large output would hang
changed: internal command execution to be faster