Add Commits Iterator interface to GitShell
I did this so that we could implement features in the future that
utilize the Sequence/Iterator approach in a more efficient fashion than
fetching and parsing all the commits up front into objects contained in
an array.

Note: I am not including a changelog entry here because this is all
internal and doesn't impact the public facing interface of this
application.