Fix a bug in commits iterator implementation
I did this because it was accidentally processing the initial commit
twice. This was happening because of a missed guard statement that I
added in this commit.

I also removed a conditional that is no longer necessary thanks to the
new guard statement.

Note: I am not including a changelog entry because this isn't currently
impacting public user experience of this application.