Add sha logging for review branch
I did this because I'd often have to go to github to see what sha the
branch was made on so I could use it in packages in swift. This saves me
a step and prints out the sha for the branch that was just checked out.

[changelog]
added: logging for the request review branch sha-1

ps-id: E8E4C2E4-FEEF-46D5-A68C-8F6EEFB029E2

Testing

Switch to HEAD

use branchname

Move the print

Move it to existing log

Add git rev parse

Added git rev-parse because it does what we want more closely.

ps-id: 6F68CA7F-EBC8-40FF-91B0-959D175C05A1