Add sha logging for review branch
I did this because I'd often have to go to github to see what sha the
branch was made on so I could use it in packages in swift. This saves me
a step and prints out the sha for the branch that was just checked out.

[changelog]
added: logging for the request review branch sha-1

ps-id: E8E4C2E4-FEEF-46D5-A68C-8F6EEFB029E2

Testing