Fix bug where git ps pub -f looked for a value
In the git ps pub cli parser it was incorrectly telling the algorithm
that the most recently identified option was "-n" when in reality it was
"-f". This was causing the parser to look for a value when it shouldn't
have.

[changelog]
fixed: bug with git ps pub -f cli parsing

ps-id: 25E2D41E-D68C-42DE-968B-157A35A100DE