Add request review of series of patches
This enables the user to invoke git ps via `git ps rr 2-4` or `git ps rr
2 4` where 2 and 4 define an inclusive series of patches via the patch
indexes 2 and 4.

If you had the following patch stack

3 4222az even more things
2 4423ab yet another thing
1 23auab Some other thing
0 auo23u Some thing

git ps rr 1-2

would request review of a patch series consisting of patch 1 & 2.

If none of the patches in the patch series have a branch already
associated with them then we grab the first patch in the series and
generate a branch name from that patch and associate that first patch to
the branch.

In the case that one of the patches in the series is already associated
to a branch, that branch will be used.

In the case that more than one patch in the series is already associated
to a branch. We error out notifying the users of the situation and
require that they explicitly specify a branch name with the `-n
<branch-name>` option.

[changelog]
added: ability to request review of a series of patches