Make commitAmendMessages() bypass commit hooks
I did this because it is something where if users have previously
explicitly bypassed commit hooks when creating the patch and then we try
to commitAmendMessages(), say to add a ps-id. The user can experience
failures from the commit hooks.

Taking this stance makes sense as the commit hooks will have already
been applied if the user wanted when they initially did the `git commit`
to create the patch. If they user explicitly bypassed the commit hooks
when creating the patch this will continue to work as it will bypass the
commit hooks as well.

This addresses issue #33.

[changelog]
changed: to bypass commit hooks when modifying commit messages

ps-id: 39754281-455C-40FD-B255-D0B21427C562