Add getShowNoColor(ref:) to GitShell
So that we can use it in the future to implement a better `rr+`
change tracking solution.

The idea is that we will compute a sha of the output of
getShowNoColor(ref:) after stripping out lines that start with `parent
`, `commit `, `tree `, or `index `.  This should give us a hash value we
can store when request for review happens and that we can recompute and
compare against that stored hash value to detect if the content of the
patch or the attributes other than tree or parent have changed. Mapping
this to the `rr+` indicator should result in a much more meaningful &
useful `rr+`.

This relates to issue #47.

ps-id: E27A224A-B330-4213-BA77-B63079E69FDA