Add String.sha1 to facilitate getting the sha1
So that in the future we can use this to implement a better version of
the `rr+` status so that it is actually meaningful. We need this to be
able to compute the sha1 of the location agnostic attributes and
content. Which in turn will allow us to compare the computed hash to one
that we store when a request for review happens.

This relates to issue #47.

ps-id: 0BA6D118-9385-4B5B-91D1-C9C6D6D5561D