Fix `rr+` state indication
Previously when you would request a review it would store a record of
the commit sha when the request for review was performed and then when
we would do a `git ps ls` it would compare the stored commit sha for the
patch against the current patch's commit sha. This didn't actually
indicate what we wanted as we really wanted `rr+` to indicate if the
diff or attributes of the patch has been modified since the request for
review was last made.

This changes things so that when we request review of a patch it
computes a special locationAgnosticHash that is stored in association
with the patch in the request review repository. It also modifies the
logic for determining the requestReviewStatus of a patch by now checking
the stored locationAgnosticHash against a currently computed location
agnostic hash. If they match then we know the patch hasn't changed since
the last request for review of it. If they don't match then we know that
the patch has changed since the last request for review. This change in
logic maintains the old logic as a fallback if the stored record for a
patch doesn't include a location agnostic hash. The stored record will
be upgraded to include the locationAgnosticHash the next time that patch
is requested for review at which point it will then use the new logic.

This resolves issue #47.

[changelog]
fixed: `rr+` statuses to properly indicate change after request review

ps-id: 646CBE00-59D9-4A7C-BC10-4D2F854FE2F3