Fix bug where rr+ status would show up when GPG sig changes
This was happening because because the output of getShow(ref:) included
the GPG signature in the output and it wasn't filtered out. This caused
the rr+ status to appear incorrectly as the GPG signature would change
when rebasing things.

This patch resolves this issue by changing the getShow(ref:) call to
specify a format of only the output we want to use to generate the hash.
This gets us 99% of the way there. However, there isn't a way with the
show command to eliminate the `index ...` line out of the patch diff
output. So we still have to filter that line out.

This is a much more stable implementation long term as we shouldn't have
to worry about unknown items popping up in the hash content like
happened with the GPG sig use case.

[changelog]
fixed: bug where rr+ status was incorrectly shown rebasing with GPG signatures

ps-id: BC34FE1D-86F1-4CA7-8817-33BACD7904F6