On rr failure cleanup ps/rr branch
This modifies the request review of single patch and request review of
series of a patch to delete the ps/rr/ branch it created in the case of
failure.

It also modifies it so that it checks out the original branch
in the case of failure. Previously it would confusingly leave the user
in the ps/rr/ branch.

Hopefully this helps button up the user experience and leave the user in
a better state when for example it fails to request review because it
couldn't apply a cherry pick, etc.

[changelog]
changed: rr to cleanup created branch & checkout original branch on failure

ps-id: CE022D41-608E-4265-AE50-F8C457A4DA71